restock/inventory_ml/requirements.txt
pymongo==4.6.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.3.2
python-dotenv==1.0.1
matplotlib==3.8.4
